requirements.txt
#git+https://github.com/nvie/rq.git#egg=rq
rq
rq-scheduler
Flask
Flask-RESTful
Flask-RQ2
flask-env-settings
gunicorn
pylokit
python-libmagic
wand
ujson
nose
